chore: JSDocs and seperating commands from workflow script
